Added function removeAccents()
I propose adding this function to remove diacritical marks.

WARNING: I'm not that versed in Java.  I just copied this off the internet but it seems like plausibly good code.  Unfortunately, it's not clear to me how to build this in my own environment and test it.

The web page that is referenced does have some text code to show how it works.

I'm creating this PR in response to: https://community.neo4j.com/t/filtering-non-english-characters-in-neo4j/30108/2

Sorry I can't be more helpful....  I know just enough to be dangerous!