Update rename-label-type-property.adoc
Initially, I was confused by committedOperations which was not described on this page.  I added documentation for it and a pointer to other values these procedures could return, plus a link to info about the YIELD key word since newbie users might not know about YIELD

I also rephrased "Procedure set" to "Set of Procedures" as I thought the former was potentially ambiguous.

I made word config into a code format so that it is easier to spot and realize it refers to the parameter in the CALLs

I added a comment for collect(X) : "// makes a list of X" because I suspect a lot of newbies who just want to rename their labels might not know about collect().  This page could be the first time they see collect()!

I changed the table column from "name" to "config parameter name". The first time I skimmed the page, I saw the "name" but I didn't immediately realize it was for the config parameters.